Rename config controller and create caches controller
